added some sanity checking

+++ all.bash
@@ -1,6 +1,18 @@
 #!/bin/bash
-go install ws
+CWD=$(pwd)
+go build
+if [ -f ws ]; then
+   mv ws $GOPATH/bin/
+else
+   echo "Something went wrong building ws."
+   exit 1
+fi
 cd extra
-go build ws-gencert.go
+go build
+if [ -f ws-genecert ]; then
 mv ws-gencert $GOPATH/bin/
-cd ..
+else
+    echo "Something went wrong building ws-gencert."
+    exit 1
+fi
+cd $CWD
